Move some utils from wallets to address & mnemonic scope
