+++ src/utils/address/index.js
@@ -1,0 +1,27 @@
+// @flow
+
+import add0x from './add0x'
+import strip0x from './strip0x'
+import checkAddressValid from './checkAddressValid'
+import getAddressChecksum from './getAddressChecksum'
+import getShortenedAddress from './getShortenedAddress'
+import checkPrivateKeyValid from './checkPrivateKeyValid'
+import getAddressFromKeyPair from './getAddressFromKeyPair'
+import getAddressWithChecksum from './getAddressWithChecksum'
+import getAddressFromPublicKey from './getAddressFromPublicKey'
+import getAddressFromPrivateKey from './getAddressFromPrivateKey'
+import checkAddressWithChecksumValid from './checkAddressWithChecksumValid'
+
+export {
+  add0x,
+  strip0x,
+  checkAddressValid,
+  getAddressChecksum,
+  getShortenedAddress,
+  checkPrivateKeyValid,
+  getAddressFromKeyPair,
+  getAddressWithChecksum,
+  getAddressFromPublicKey,
+  getAddressFromPrivateKey,
+  checkAddressWithChecksumValid,
+}
