+++ src/utils/mnemonic/checkDerivationPathValid.js
