+++ src/utils/mnemonic/checkMnemonicValid.js
