+++ src/utils/mnemonic/generateAddresses.js
