+++ src/utils/mnemonic/getMnemonicOptions.js
