+++ src/utils/wallets/getWallet.js
@@ -2,13 +2,13 @@
 
 function getWallet(wallets: Wallets, walletId: ?string): Wallet {
   if (!walletId) {
-    throw new Error('Wallet ID not provided')
+    throw new Error('WalletNotFoundError')
   }
 
   const wallet: ?Wallet = wallets.find(({ id }: Wallet): boolean => (walletId === id))
 
   if (!wallet) {
-    throw new Error(`Wallet with id ${walletId} not found`)
+    throw new Error(`WalletNotFoundError ${walletId}`)
   }
 
   return { ...wallet }
