+++ src/utils/wallets/index.js
@@ -1,8 +1,5 @@
 // @flow
 
-import add0x from './add0x'
-import strip0x from './strip0x'
-import getHdPath from './getHdPath'
 import getWallet from './getWallet'
 import getAddress from './getAddress'
 import getMnemonic from './getMnemonic'
@@ -14,33 +11,10 @@ import getAddresses from './getAddresses'
 import getBackupData from './getBackupData'
 import getPrivateKey from './getPrivateKey'
 import getTypeByInput from './getTypeByInput'
-import generateAddress from './generateAddress'
-import getPublicHdRoot from './getPublicHdRoot'
-import generateMnemonic from './generateMnemonic'
-import getPrivateHdRoot from './getPrivateHdRoot'
-import checkAddressValid from './checkAddressValid'
 import checkMnemonicType from './checkMnemonicType'
-import generateAddresses from './generateAddresses'
-import checkMnemonicValid from './checkMnemonicValid'
-import getAddressChecksum from './getAddressChecksum'
-import getMnemonicOptions from './getMnemonicOptions'
-import getShortenedAddress from './getShortenedAddress'
-import getXPubFromMnemonic from './getXPubFromMnemonic'
-import checkPrivateKeyValid from './checkPrivateKeyValid'
 import checkWalletUniqueness from './checkWalletUniqueness'
-import getAddressFromKeyPair from './getAddressFromKeyPair'
-import getAddressWithChecksum from './getAddressWithChecksum'
-import getAddressFromPublicKey from './getAddressFromPublicKey'
-import checkDerivationPathValid from './checkDerivationPathValid'
-import getAddressFromPrivateKey from './getAddressFromPrivateKey'
-import checkBip32XPublicKeyValid from './checkBip32XPublicKeyValid'
-import getPrivateKeyFromMnemonic from './getPrivateKeyFromMnemonic'
-import checkAddressWithChecksumValid from './checkAddressWithChecksumValid'
 
 export {
-  add0x,
-  strip0x,
-  getHdPath,
   getWallet,
   getAddress,
   getMnemonic,
@@ -52,26 +26,6 @@ export {
   getBackupData,
   getPrivateKey,
   getTypeByInput,
-  generateAddress,
-  getPublicHdRoot,
-  generateMnemonic,
-  getPrivateHdRoot,
-  checkAddressValid,
   checkMnemonicType,
-  generateAddresses,
-  checkMnemonicValid,
-  getAddressChecksum,
-  getMnemonicOptions,
-  getShortenedAddress,
-  getXPubFromMnemonic,
-  checkPrivateKeyValid,
   checkWalletUniqueness,
-  getAddressFromKeyPair,
-  getAddressWithChecksum,
-  getAddressFromPublicKey,
-  checkDerivationPathValid,
-  getAddressFromPrivateKey,
-  checkBip32XPublicKeyValid,
-  getPrivateKeyFromMnemonic,
-  checkAddressWithChecksumValid,
 }
